requirements.txt
pydantic>=2.0.0,<3.0.0
numpy>=1.24.0
pandas>=2.0.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
